[#540] Refactor require statements to use 'app_settings' instead of 'base' for configuration settings
